Lowered ram usage, fixed terminal
Switched datatypes to bytes,
Used flash memory for printing strings

Fixed problem with split string returning whole word
Rewrote typos in prints
Fixed randomize (also ruint randomize)

+++ Ktane_Code/Joystick.cpp
@@ -10,7 +10,7 @@ class Joystick {
 };
 class Joystick_TwoWay : public Joystick {
   private:
-    int list[7] = {B110, B100, B101, B001, B011, B010, B000};
+    byte list[7] = {B110, B100, B101, B001, B011, B010, B000};
   public:
     Joystick_TwoWay(int pin1, int pin2, int pin3) {
       _pin1 = pin1;
@@ -19,7 +19,7 @@ class Joystick_TwoWay : public Joystick {
 
     }
 
-    int getPos() {
+    byte getPos() {
       currVal = digitalRead(_pin1) * 4 + digitalRead(_pin2) * 2 + digitalRead(_pin3);
       for (int l = 0; l < 7; l++) {
         if (list[l] == currVal)
@@ -27,7 +27,7 @@ class Joystick_TwoWay : public Joystick {
       }
       return 0;
     }
-    int getRaw() {
+    byte getRaw() {
       currVal = digitalRead(_pin1) * 4 + digitalRead(_pin2) * 2 + digitalRead(_pin3);
       return currVal;
     }
@@ -35,7 +35,7 @@ class Joystick_TwoWay : public Joystick {
 
 class Joystick_Lever : public Joystick {
   private:
-    int list[15] = {B0010, B0000, B1000, B1010, B1110, B1100, B1101, B1001, B1011, B0011, B0001, B0101, B0100, B0110, B0111};
+    byte list[15] = {B0010, B0000, B1000, B1010, B1110, B1100, B1101, B1001, B1011, B0011, B0001, B0101, B0100, B0110, B0111};
   public:
     Joystick_Lever(int pin1, int pin2, int pin3, int pin4) {
       _pin1 = pin1;
@@ -44,7 +44,7 @@ class Joystick_Lever : public Joystick {
       _pin4 = pin4;
     }
 
-    int getPos() {
+    byte getPos() {
       currVal = digitalRead(_pin1) * 8 + digitalRead(_pin2) * 4 + digitalRead(_pin3) * 2 + digitalRead(_pin4);
       for (int l = 0; l < 15; l++) {
         if (list[l] == currVal)
@@ -52,7 +52,7 @@ class Joystick_Lever : public Joystick {
       }
       return 0;
     }
-    int getRaw() {
+    byte getRaw() {
       currVal = digitalRead(_pin1) * 8 + digitalRead(_pin2) * 4 + digitalRead(_pin3) * 2 + digitalRead(_pin4);
       return currVal;
     }
